Revised structure, css and js.
Changed a bit the approach to fit with the need for the code tutorial part.
