+++ general_guidance/index.html
@@ -5,22 +5,32 @@
   <meta charset="UTF-8" />
   <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <link href="/assets/css/main.min.css" rel="stylesheet" />
-  <title>UNHCR Data Visualization Platform</title>
-  <!-- TODO change icon link with final URL -->
-  <link rel="shortcut icon" href="http://127.0.0.1:5500//favicon.ico">
+  <meta name="author" content="IPDA/GDS">
+  <title>UNHCR Dataviz - General guidance</title>
+  <link href="/assets/css/unhcrmain.min.css" rel="stylesheet" />
+  <!-- prism CSS file -->
+  <link href="/assets/css/vendor/prism.min.css" rel="stylesheet" />
+  <!-- prism JS file -->
+  <script src="/assets/js/vendor/prism.min.js"></script>
+  <!-- jquery JS file -->
+  <script src="/assets/js/vendor/jquery.min.js"></script>
+  <!-- d3 JS file -->
+  <script src="/assets/js/vendor/d3.min.js"></script>
+  <!-- parser JS file -->
+  <script src="/assets/js/vendor/parser.js"></script>
 </head>
 
-<body class="d-flex flex-column vh-100" id="general_guidance">
+<body class="d-flex flex-column vh-100">
   <!-- Navigation -->
   <hcr-header></hcr-header>
   <!-- End Navigation -->
 
   <!-- Main -->
-  <main class="flex-shrink-0">
+  <main id="general_guidance" class="d-flex flex-column min-vh-100 flex-shrink-0">
     <div class="container my-5">
-      <div class="section-title mb-5 text-center">
-        <h2>General guidance</h2>
+      <div class="text-center mb-5">
+        <h2 class="text-uppercase">General guidance</h2>
+        <hr class="short-hr">
         <div class="row justify-content-center">
           <div class="col-lg-8">
             <p>Magnam dolores commodi suscipit. Necessitatibus eius consequatur ex aliquid fuga eum quidem. Sit sint
@@ -30,7 +40,8 @@
           </div>
         </div>
       </div>
-      <section>
+
+      <section class="my-5">
         <h3>Data Visualization Guidelines</h3>
         <p>Data-driven products help us communicate the relevance and impact of UNHCR’s work to the public. All charts
           and data visualizations created for UNHCR should follow the Data Visualization Guidelines that you can
@@ -451,17 +462,17 @@
   <hcr-footer></hcr-footer>
   <!-- End footer -->
 
-  <!-- JavaScript Bundle with Popper -->
-  <script src="/assets/js/bootstrap.bundle.min.js"></script>
+  <!-- Scripts -->
+  <!-- bs/popper bundle -->
+  <script src="/assets/js/vendor/bootstrap.bundle.min.js"></script>
+  <!-- UNHCR includes header/footer JS file -->
+  <script type="module" src="/assets/js/unhcrincludes.js"></script>
+  <!-- UNHCR main JS file -->
+  <script type="module" src="/assets/js/unhcrmain.js"></script>
 
-  <!-- Main JS file -->
-  <script type="module" src="/assets/js/main.js"></script>
+  <script src="/assets/js/vendor/isotope.pkgd.min.js"></script>
+  <!-- End scripts -->
 
-  <!-- Vendor JS file -->
-  <script src="/assets/js/jquery.min.js"></script>
-  <!-- <script src="assets/js/jquery.easing.min.js"></script> -->
-  <script src="/assets/js/isotope.pkgd.min.js"></script>
-  <script src="/assets/js/aos.js"></script>
 </body>
 
 </html>
