+++ index.html
@@ -5,26 +5,35 @@
   <meta charset="UTF-8" />
   <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <link href="/assets/css/main.min.css" rel="stylesheet" />
-  <title>UNHCR Data Visualization Platform</title>
-  <!-- TODO change icon link with final URL -->
-  <link rel="shortcut icon" href="http://127.0.0.1:5500//favicon.ico">
+  <meta name="author" content="IPDA/GDS">
+  <title>UNHCR Dataviz</title>
+  <link href="/assets/css/unhcrmain.min.css" rel="stylesheet" />
+  <!-- prism CSS file -->
+  <link href="/assets/css/vendor/prism.min.css" rel="stylesheet" />
+  <!-- prism JS file -->
+  <script src="/assets/js/vendor/prism.min.js"></script>
+  <!-- jquery JS file -->
+  <script src="/assets/js/vendor/jquery.min.js"></script>
+  <!-- d3 JS file -->
+  <script src="/assets/js/vendor/d3.min.js"></script>
+  <!-- parser JS file -->
+  <script src="/assets/js/vendor/parser.js"></script>
 </head>
 
-<body class="d-flex flex-column vh-100" id="home">
+<body class="d-flex flex-column vh-100">
   <!-- Navigation -->
   <hcr-header></hcr-header>
   <!-- End Navigation -->
 
   <!-- Main -->
-  <main class="flex-shrink-0">
-    <section class="banner py-5 mb-5">
+  <main id="home" class="flex-shrink-0">
+    <section class="banner py-5" style="background-image: url(/assets/img/home_banner.svg); background-size:cover; background-position: center; background-repeat: no-repeat;">
       <div class="container py-5">
-        <h1 class="main-title">Data visualization platform</h1>
-        <p class="tagline">Insights, guidelines, and tools for better data storytelling.</p>
+        <h1 class="display-4 fw-bold">Data visualization platform</h1>
+        <p class="fs-4">Insights, guidelines, and tools for better data storytelling.</p>
       </div>
     </section>
-    <div class="container mb-5">
+    <div class="container my-5">
       <div class="row row-cols-1 row-cols-md-2 g-5">
         <div class="col">
           <a href="/chart_gallery/">
@@ -66,17 +75,17 @@
   <hcr-footer></hcr-footer>
   <!-- End footer -->
 
-  <!-- JavaScript Bundle with Popper -->
-  <script src="/assets/js/bootstrap.bundle.min.js"></script>
+  <!-- Scripts -->
+  <!-- bs/popper bundle -->
+  <script src="/assets/js/vendor/bootstrap.bundle.min.js"></script>
+  <!-- UNHCR includes header/footer JS file -->
+  <script type="module" src="/assets/js/unhcrincludes.js"></script>
+  <!-- UNHCR main JS file -->
+  <script type="module" src="/assets/js/unhcrmain.js"></script>
 
-  <!-- Main JS file -->
-  <script type="module" src="/assets/js/main.js"></script>
+  <script src="/assets/js/vendor/isotope.pkgd.min.js"></script>
+  <!-- End scripts -->
 
-  <!-- Vendor JS file -->
-  <script src="/assets/js/jquery.min.js"></script>
-  <!-- <script src="assets/js/jquery.easing.min.js"></script> -->
-  <script src="/assets/js/isotope.pkgd.min.js"></script>
-  <script src="/assets/js/aos.js"></script>
 </body>
 
 </html>
