--- sass/_unhcrcustom.scss
@@ -1,309 +1,0 @@
-// General
-body{
-  position: relative;
-}
-.container {
-  max-width: 1200px;
-}
-.bg-navy {
-    background-color: $unhcr-navy;
-}
-.bg-yellow {
-  background-color: $unhcr-yellow;
-}
-
-// Typography
-li::marker {
-  color: $unhcr-blue;
-}
-
-// Button
-.btn-primary:hover,
-.btn-primary:active,
-.btn-primary:focus {
-  color: $unhcr-white;
-  background-color: $unhcr-darkgrey;
-  border-color: $unhcr-white;
-}
-
-// Header
-.navbar {
-  font-size: 0.875rem;
-  letter-spacing: 0.075rem;
-
-  .navbar-brand img {
-    height: 3rem;
-    max-height: 2.75rem;
-  }
-  .navbar-toggler {
-    border: none;
-    .navbar-toggler-icon {
-      background-image: url("data:image/svg+xml;charset=utf8,%3Csvg viewBox='0 0 32 32' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath stroke='rgba(0,114,188,1)' stroke-width='2' stroke-linecap='round' stroke-miterlimit='10' d='M4 8h24M4 16h24M4 24h24'/%3E%3C/svg%3E");
-    }
-  }
-  .navbar-toggler:focus,
-  .navbar-toggler:active,
-  .navbar-toggler-icon:focus {
-    box-shadow: none;
-  }
-  .nav-item {
-    .nav-link {
-      color: $unhcr-grey;
-    }
-    .nav-link:hover,
-    .nav-link.active {
-      color: $unhcr-blue;
-      border-bottom: 2pt solid $unhcr-blue;
-    }
-  }
-}
-// To uncomment if we add dropdown
-// .dropdown-menu {
-//   background-color: $gray-100;
-//   border: none;
-//   a:hover {
-//     color: $unhcr-white;
-//     background-color: $unhcr-blue;
-//   }
-//   .dropdown-item {
-//     font-size: 0.8rem;
-//   }
-// }
-// .dropdown:hover .dropdown-menu {
-//   display: block;
-// }
-
-@media (min-width: 992px) {
-  .navbar-expand-lg .navbar-nav .nav-link {
-    padding: 0.25rem 0;
-    margin: 0.25rem 0.75rem;
-  }
-}
-
-// Footer
-footer {
-  .nav {
-    font-size: 0.875rem;
-  }
-  .nav-link {
-    color: $gray-100;
-  }
-  .nav-link:hover {
-    color: $gray-400;
-  }
-}
-
-// Home banner
-.banner {
-  background-image: url(/assets/img/home_banner.svg);
-  background-color: $unhcr-mediumgrey;
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: cover;
-}
-.main-title {
-  font-size: 3rem;
-  font-weight: 700;
-}
-.tagline {
-  font-size: 1.5rem;
-  font-weight: 300;
-}
-
-// Main page title
-.section-title {
-  h2 {
-    text-transform: uppercase;
-    font-weight: 700 !important;
-    position: relative;
-    margin-bottom: 1rem;
-    padding-bottom: 0.75rem;
-  }
-  h2::after {
-    content: "";
-    position: absolute;
-    display: block;
-    width: 3rem;
-    height: 0.2rem;
-    background: $unhcr-blue;
-    bottom: 0;
-    left: calc(50% - 1.5rem);
-  }
-}
-
-// Chart gallery
-#chart {
-  overflow: hidden;
-  .chart-menu {
-    margin-bottom: 40px;
-    li {
-      cursor: pointer;
-      display: inline-block;
-      list-style: none;
-      margin: 0;
-      font-size: 15px;
-      font-weight: 500;
-      line-height: 1;
-      color: $gray-800;
-      transition: all 0.3s;
-      padding: 8px 10px;
-    }
-    li:hover,
-    li.filter-active {
-      background: $unhcr-blue;
-      color: $unhcr-white;
-    }
-  }
-  .chart-item {
-    margin-bottom: 2rem;
-    .chart-img {
-      overflow: hidden;
-      transition: 0.3s;
-      position: relative;
-      z-index: 1;
-      img {
-        transition: all 0.6s;
-      }
-    }
-    .chart-info {
-      opacity: 0;
-      position: absolute;
-      top: 0;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      z-index: 3;
-      transition: all ease-in-out 0.5s;
-      background: rgba(35, 35, 35, 0.6);
-      padding: 0.6rem;
-      .btn {
-        font-size: 14px;
-        color: $unhcr-white;
-        background: none;
-      }
-    }
-  }
-  .chart-item:hover .chart-img img {
-    transform: scale(1.1);
-  }
-  .chart-item:hover .chart-info {
-    opacity: 1;
-  }
-}
-
-// Chart page with sidebar
-.sidebar-area{
-  margin: 0px;
-  .nav-pills{
-    margin-left: 0em;
-    .navbar-title{
-      color: $gray-800;
-      padding: 0.25rem 0.5rem;
-    }
-    .nav-link{
-      font-size: small;
-      color:$unhcr-grey;
-      padding: 0.25rem 0.5rem;
-    }
-    .nav-link.active,
-    .nav-link:hover,
-    .show>.nav-link {
-    background-color: $gray-200;
-    color: $unhcr-blue;
-    }
-  }
-}
-
-.chart-area{
-  margin-bottom: 5rem;
-  .chart-img img {
-    width: 10rem;
-    padding-top: 5rem;
-    padding-bottom: 1.5rem;
-  }
-  h2{
-    margin-bottom: 5rem;
-  }
-  h4{
-    margin-top: 2rem;
-  }
-  .card{
-    border: none;
-    margin-top: 1rem;
-    .card-body{
-      padding:0.5rem 0rem 1rem;
-      p{
-        color: $gray-600;
-      }
-    }
-  }
-}
-
-hr{
-  margin-top: 5rem;
-}
-
-//Back to top button
-#button {
-  display: inline-block;
-  background-color: $unhcr-grey;
-  width: 2.5rem;
-  height: 2.5rem;
-  text-align: center;
-  position: fixed;
-  bottom: 8rem;
-  right: 1rem;
-  transition: background-color .3s, 
-    opacity .5s, visibility .5s;
-  opacity: 0;
-  visibility: hidden;
-  z-index: 1000;
-}
-@import url(https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css);
-#button::after {
-  content: "\f077";
-  font-family: FontAwesome;
-  font-weight:400;
-  font-style: normal;
-  font-size: 1.5em;
-  line-height: 2.2rem;
-  color: $unhcr-white;
-}
-#button:hover,
-#button:active {
-  cursor: pointer;
-  background-color: $unhcr-blue;
-  text-decoration: none;
-}
-#button.show {
-  opacity: 0.5;
-  visibility: visible;
-}
-
-//Product page filter
-#product {
-  overflow: hidden;
-  .product-menu {
-    margin-bottom: 40px;
-    li {
-      cursor: pointer;
-      display: inline-block;
-      list-style: none;
-      margin: 0;
-      font-size: 15px;
-      font-weight: 500;
-      line-height: 1;
-      color: $gray-800;
-      transition: all 0.3s;
-      padding: 8px 10px;
-    }
-    li:hover,
-    li.filter-active {
-      background: $unhcr-blue;
-      color: $unhcr-white;
-    }
-  }
-}
-
-
-
