+++ sass/_unhcrfilters.scss
@@ -1,0 +1,54 @@
+// Filter button
+#list-filter {
+  li {
+    cursor: pointer;
+    transition: all 0.3s;
+    padding: 0.25rem 0.5rem;
+    transition: all 0.3s;
+  }
+  li:hover,
+  li.filter-active {
+    background: $unhcr-blue;
+    color: $unhcr-white;
+  }
+}
+
+// Chart icon pages
+.chart-item {
+  .chart-img {
+    overflow: hidden;
+    transition: 0.3s;
+    position: relative;
+    z-index: 1;
+    img {
+      transition: all 0.6s;
+    }
+  }
+  .chart-info {
+    opacity: 1;
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    z-index: 3;
+    transition: all ease-in-out 0.5s;
+    background: rgba(35, 35, 35, 0.6);
+    .chart-name {
+      text-align: start;
+      line-height: 1.2;
+    }
+  }
+}
+
+@media (min-width: 576px) {
+  .chart-item .chart-info {
+    opacity: 0;
+  }
+  .chart-item:hover .chart-img img {
+    transform: scale(1.1);
+  }
+  .chart-item:hover .chart-info {
+    opacity: 1;
+  }
+}
