+++ sass/_unhcrfooter.scss
@@ -1,0 +1,11 @@
+footer {
+  .nav {
+    font-size: 0.875rem;
+  }
+  .nav-link {
+    color: $gray-100;
+  }
+  .nav-link:hover {
+    color: $gray-400;
+  }
+}
