+++ sass/_unhcrnavbar.scss
@@ -1,0 +1,52 @@
+.navbar {
+  font-size: 0.875rem;
+  letter-spacing: 0.075rem;
+
+  .navbar-brand img {
+    height: 3rem;
+    max-height: 2.75rem;
+  }
+  .navbar-toggler {
+    border: none;
+    .navbar-toggler-icon {
+      background-image: url("data:image/svg+xml;charset=utf8,%3Csvg viewBox='0 0 32 32' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath stroke='rgba(0,114,188,1)' stroke-width='2' stroke-linecap='round' stroke-miterlimit='10' d='M4 8h24M4 16h24M4 24h24'/%3E%3C/svg%3E");
+    }
+  }
+  .navbar-toggler:focus,
+  .navbar-toggler:active,
+  .navbar-toggler-icon:focus {
+    box-shadow: none;
+  }
+  .nav-item {
+    .nav-link {
+      color: $unhcr-grey;
+    }
+    .nav-link:hover,
+    .nav-link.active {
+      color: $unhcr-blue;
+      border-bottom: 2pt solid $unhcr-blue;
+    }
+  }
+}
+// To uncomment if we add dropdown
+// .dropdown-menu {
+//   background-color: $gray-100;
+//   border: none;
+//   a:hover {
+//     color: $unhcr-white;
+//     background-color: $unhcr-blue;
+//   }
+//   .dropdown-item {
+//     font-size: 0.8rem;
+//   }
+// }
+// .dropdown:hover .dropdown-menu {
+//   display: block;
+// }
+
+@media (min-width: 992px) {
+  .navbar-expand-lg .navbar-nav .nav-link {
+    padding: 0.25rem 0;
+    margin: 0.25rem 0.75rem;
+  }
+}
