+++ sass/unhcrmain.scss
@@ -61,4 +61,8 @@ $colors: map-merge($colors, $custom-colors);
 // @import "../node_modules/bootstrap/scss/utilities/api";
 
 // 6. Add additional custom code here
-@import "unhcrcustom";
+@import "unhcrglobal";
+@import "unhcrnavbar";
+@import "unhcrfooter";
+@import "unhcrfilters";
+@import "unhcrsidebar";
